requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
cryptography==44.0.2
Django==5.1.7
django-storages==1.14.5
djangorestframework==3.16.0
gunicorn==23.0.0
idna==2.10
isodate==0.7.2
packaging==24.2
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.1.0
requests==2.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==1.26.20
